small_django_project/feeds/api/serializers.py
from rest_framework import serializers

from feeds.models import FeedItem


class FeedItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = FeedItem
        fields = ('id', 'content', 'user')
